Fixed wrong reference in test
