+++ app/src/main/java/com/eveningoutpost/dexdrip/utils/framework/RetrofitService.java
@@ -5,6 +5,7 @@ import static com.eveningoutpost.dexdrip.utilitymodels.OkHttpWrapper.enableTls12
 
 import com.eveningoutpost.dexdrip.cgm.nsfollow.GzipRequestInterceptor;
 import com.eveningoutpost.dexdrip.models.UserError;
+import com.eveningoutpost.dexdrip.store.FastStore;
 import com.eveningoutpost.dexdrip.tidepool.InfoInterceptor;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -20,8 +21,12 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
+import lombok.RequiredArgsConstructor;
 import okhttp3.OkHttpClient;
 import okhttp3.logging.HttpLoggingInterceptor;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
@@ -60,7 +65,7 @@ public class RetrofitService {
         final Gson gson = new GsonBuilder()
                 .registerTypeAdapterFactory(UNRELIABLE_INTEGER_FACTORY)
                 .create();
-        Retrofit retrofit = new retrofit2.Retrofit.Builder()
+        Retrofit retrofit = new Retrofit.Builder()
                 .baseUrl(url)
                 .client(client)
                 .addConverterFactory(GsonConverterFactory.create(gson))
@@ -81,6 +86,65 @@ public class RetrofitService {
 
     public static int size() {
         return RETROFIT_CACHE.size();
+    }
+
+    // ===== Boilerplate classes ===================================================================
+    // Callback template to reduce boiler plate
+
+    @RequiredArgsConstructor
+    public static  class BaseCallback<T> implements Callback<T> {
+
+        protected final String TAG = this.getClass().getSimpleName();
+
+        //  final Session session;
+        protected final String name;
+        private Runnable onSuccess;
+        private Runnable onFailure;
+
+        public BaseCallback<T> setOnSuccess(final Runnable runnable) {
+            this.onSuccess = runnable;
+            return this;
+        }
+
+        public BaseCallback<T> setOnFailure(final Runnable runnable) {
+            this.onFailure = runnable;
+            return this;
+        }
+
+        @Override
+        public void onResponse(Call<T> call, Response<T> response) {
+            if (response.isSuccessful() && response.body() != null) {
+                UserError.Log.d(TAG, name + " success");
+                if (onSuccess != null) {
+                    onSuccess.run();
+                }
+            } else {
+                final String msg = name + " was not successful: " + response.code() + " " + response.message();
+                UserError.Log.e(TAG, msg);
+                status(msg);
+                if (onFailure != null) {
+                    onFailure.run();
+                }
+            }
+        }
+
+        @Override
+        public void onFailure(Call<T> call, Throwable t) {
+            final String msg = name + " Failed: " + t;
+            UserError.Log.e(TAG, msg);
+            status(msg);
+            if (onFailure != null) {
+                onFailure.run();
+            }
+        }
+
+        public String getStatus() {
+            return FastStore.getInstance().getS(TAG + ".STATUS_KEY");
+        }
+
+        private void status(final String status) {
+            FastStore.getInstance().putS(TAG + ".STATUS_KEY", status);
+        }
     }
 
     // ===== Private helpers =======================================================================
@@ -120,6 +184,5 @@ public class RetrofitService {
             out.value(value);
         }
     };
-    private static final TypeAdapterFactory UNRELIABLE_INTEGER_FACTORY = TypeAdapters.newFactory(int.class, Integer.class, UNRELIABLE_INTEGER);
-
+    public static final TypeAdapterFactory UNRELIABLE_INTEGER_FACTORY = TypeAdapters.newFactory(int.class, Integer.class, UNRELIABLE_INTEGER);
 }
